Move search tests into own subfolder
