+++ tests/search/search_fixtures/vespa_test_schema/schemas/document_passage.sd
