+++ tests/search/search_fixtures/vespa_test_schema/schemas/search_weights.sd
