+++ tests/search/test_vespasearch.py
@@ -12,7 +12,7 @@ from sqlalchemy.orm import Session
 from app.api.api_v1.routers import search
 from app.api.api_v1.schemas import search as search_schemas
 from app.core.lookups import get_country_slug_from_country_code
-from tests.routes.setup_search_tests import (
+from tests.search.setup_search_tests import (
     VESPA_FIXTURE_COUNT,
     _create_document,
     _create_family,
